+++ src/RuleSet/Php55.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Ergebnis\PhpCsFixer\Config\RuleSet;
 
 use Ergebnis\PhpCsFixer\Config\Name;
+use Ergebnis\PhpCsFixer\Config\PhpVersion;
 
 final class Php55 extends AbstractRuleSet implements ExplicitRuleSet
 {
@@ -825,7 +826,7 @@ final class Php55 extends AbstractRuleSet implements ExplicitRuleSet
 
     public function name(): Name
     {
-        return Name::fromString('ergebnis (PHP 5.5)');
+        return Name::fromPhpVersion($this->targetPhpVersion());
     }
 
     public function rules(): array
@@ -833,8 +834,12 @@ final class Php55 extends AbstractRuleSet implements ExplicitRuleSet
         return $this->rules;
     }
 
-    public function targetPhpVersion(): int
+    public function targetPhpVersion(): PhpVersion
     {
-        return 50500;
+        return PhpVersion::create(
+            PhpVersion\Major::fromInt(5),
+            PhpVersion\Minor::fromInt(5),
+            PhpVersion\Patch::fromInt(0),
+        );
     }
 }