+++ src/RuleSet/Php71.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Ergebnis\PhpCsFixer\Config\RuleSet;
 
 use Ergebnis\PhpCsFixer\Config\Name;
+use Ergebnis\PhpCsFixer\Config\PhpVersion;
 
 final class Php71 extends AbstractRuleSet implements ExplicitRuleSet
 {
@@ -828,7 +829,7 @@ final class Php71 extends AbstractRuleSet implements ExplicitRuleSet
 
     public function name(): Name
     {
-        return Name::fromString('ergebnis (PHP 7.1)');
+        return Name::fromPhpVersion($this->targetPhpVersion());
     }
 
     public function rules(): array
@@ -836,8 +837,12 @@ final class Php71 extends AbstractRuleSet implements ExplicitRuleSet
         return $this->rules;
     }
 
-    public function targetPhpVersion(): int
+    public function targetPhpVersion(): PhpVersion
     {
-        return 70100;
+        return PhpVersion::create(
+            PhpVersion\Major::fromInt(7),
+            PhpVersion\Minor::fromInt(1),
+            PhpVersion\Patch::fromInt(0),
+        );
     }
 }