+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`5.15.1...main`][5.15.1...main].
 ### Added
 
 - Extracted `Name` as a value object ([#870]), by [@localheinz]
+- Extracted `PhpVersion` as a value object ([#871]), by [@localheinz]
 
 ### Fixed
 
@@ -1163,6 +1164,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#867]: https://github.com/ergebnis/php-cs-fixer-config/pull/867
 [#868]: https://github.com/ergebnis/php-cs-fixer-config/pull/868
 [#870]: https://github.com/ergebnis/php-cs-fixer-config/pull/870
+[#871]: https://github.com/ergebnis/php-cs-fixer-config/pull/871
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler